Added PDIRK and PRK solvers
