+++ lib/OrdinaryDiffEqPDIRK/src/alg_utils.jl
@@ -1,0 +1,3 @@
+isfsal(alg::PDIRK44) = false
+
+alg_order(alg::PDIRK44) = 4
