+++ lib/OrdinaryDiffEqPDIRK/src/algorithms.jl
@@ -1,0 +1,24 @@
+################################################################################
+
+#################################################
+"""
+PDIRK44: Parallel Diagonally Implicit Runge-Kutta Method
+A 2 processor 4th order diagonally non-adaptive implicit method.
+"""
+struct PDIRK44{CS, AD, F, F2, P, FDT, ST, CJ, TO} <:
+       OrdinaryDiffEqNewtonAlgorithm{CS, AD, FDT, ST, CJ}
+    linsolve::F
+    nlsolve::F2
+    precs::P
+    extrapolant::Symbol
+    threading::TO
+end
+function PDIRK44(; chunk_size = Val{0}(), autodiff = true, standardtag = Val{true}(),
+        concrete_jac = nothing, diff_type = Val{:forward},
+        linsolve = nothing, precs = DEFAULT_PRECS, nlsolve = NLNewton(),
+        extrapolant = :constant, threading = true)
+    PDIRK44{_unwrap_val(chunk_size), _unwrap_val(autodiff), typeof(linsolve),
+        typeof(nlsolve), typeof(precs), diff_type, _unwrap_val(standardtag),
+        _unwrap_val(concrete_jac), typeof(threading)}(linsolve, nlsolve, precs,
+        extrapolant, threading)
+end
