+++ lib/OrdinaryDiffEqPRK/src/OrdinaryDiffEqPRK.jl
@@ -1,0 +1,16 @@
+module OrdinaryDiffEqPRK
+
+import OrdinaryDiffEq: OrdinaryDiffEqAlgorithm, alg_order, OrdinaryDiffEqMutableCache,
+                       OrdinaryDiffEqConstantCache, constvalue, @unpack, @cache, alg_cache,
+                       unwrap_alg, perform_step!
+using MuladdMacro: @muladd
+import FastBroadcast: @..
+
+include("algorithms.jl")
+include("alg_utils.jl")
+include("prk_caches.jl")
+include("prk_perform_step.jl")
+
+export KuttaPRK2p5
+
+end
