+++ lib/OrdinaryDiffEqPRK/src/alg_utils.jl
@@ -1,0 +1,1 @@
+alg_order(alg::KuttaPRK2p5) = 5
